Replace Qt header by Ethereum header.
